[Refacotor] 컴포넌트 분리 및 메서드 변경
